Fix build against newer GCC and the bumped pico-sdk
None of this is specific to the VMU work; it is what the tree needs to
compile at all with a current toolchain, and it fails the same way on an
unmodified upstream checkout.

- MapleBusInterface.hpp uses std::uint64_t throughout MapleStats but
  included only <stdint.h>, relying on a transitive <cstdint> that newer
  libstdc++ no longer provides.
- WebUsbCommandHandler.hpp calls std::reverse without including
  <algorithm>, for the same reason.
- FlycastWebUsbCommandHandler and SystemWebUsbCommandHandler passed
  strlen(buffer) + 1 into a uint16_t field, which newer GCC rejects
  under -Werror=narrowing. Made the conversion explicit.

Note that configuring also needs CMAKE_POLICY_VERSION_MINIMUM=3.5 with
CMake 4.x, because the pico-sdk pioasm sub-build still declares a
cmake_minimum_required below 3.5. That is a submodule issue and is not
fixed here.

+++ src/hostLib/webusb_parsers/FlycastWebUsbCommandHandler.cpp
@@ -142,7 +142,7 @@ void FlycastWebUsbCommandHandler::process(
             char buffer[mIdentification.getSerialSize() + 1] = {0};
             mIdentification.getSerial(buffer, sizeof(buffer) - 1);
             buffer[sizeof(buffer) - 1] = '\0';
-            responseFn(kResponseSuccess, {{buffer, strlen(buffer) + 1}});
+            responseFn(kResponseSuccess, {{buffer, static_cast<uint16_t>(strlen(buffer) + 1)}});
         }
         return;
